requirements.txt
asgiref==3.4.1
attrs==21.2.0
boto3==1.20.5
botocore==1.23.5
Django==3.2.9
django-phonenumber-field==6.0.0
django-redis==5.0.0
django-storages==1.12.3
djangorestframework==3.12.4
djangorestframework-gis==0.17
drf-spectacular==0.21.0
gunicorn==20.1.0
importlib-resources==5.4.0
inflection==0.5.1
jmespath==0.10.0
jsonschema==4.2.1
phonenumberslite==8.12.37
psycopg2-binary==2.9.2
pyrsistent==0.18.0
python-dateutil==2.8.2
pytz==2021.3
PyYAML==6.0
redis==3.5.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.7
zipp==3.6.0
